♻️ Factor out policyform
